protocol: initial implementation of the text protocol MessageBuffer
This defines the MessageBuffer interface that is intended to wrap
around a TCP connection and deal with reading and writing unpacked
messages. This also includes the first pass of the text protocol's
implementation of MessageBuffer that can read the storage commands.

Tests are sparse, but more to come as the Read() implementation
is completed for retrieval commands and Write() is implemented.