protocol,server: implement a text protocol session and start up server
This creates text protocol session that wraps a tcp connection, text
message buffer, and storage engine. Its Serve() method is polled
for new receiving messages. Upon receiving a message, it talks to the
storage engine to handle it, writing the proper respons. At this point,
*any* error closes the connection and the session is ended. This is subject
to change once the specs for correctness are fully met.

Also included this commit is a simple server implementation that listens
on port 11211 and spins up a new goroutine for each session created, where
the goroutine simply polls over TextProtocolSession.Serve() while the
session is alive. The server implementation may change in favor of having
a finite set of goroutine workers.

This has been crudely tested via echo & nc. With this end to end functionality
(from connection to message buffer to session to storage engine), the
rest of the protocol logic can be easily filled in.