main,protocol,store: require and support an eviction policy in the simple storage engine
This adds a required eviction policy passed to the storage engine. Currently, only
lru eviction is supported, but other policies can be easily supported via the
EvictionPolicy interface.

You can now pass a port and cap command line args to specify the server port to
run on and the total capacity of the cache in bytes.